+++ packages/components-js/tests/e2e/specs/tabs.e2e.ts
@@ -320,11 +320,14 @@ test.describe('lifecycle', () => {
     expect(status.componentDidLoad['p-tabs'], 'componentDidLoad: p-tabs').toBe(1);
     expect(status.componentDidLoad['p-tabs-bar'], 'componentDidLoad: p-tabs-bar').toBe(1); // includes 4 didLoad calls
     expect(status.componentDidLoad['p-tabs-item'], 'componentDidLoad: p-tabs-item').toBe(3);
+    expect(status.componentDidLoad['p-scroller'], 'componentDidLoad: p-scroller').toBe(1);
+    expect(status.componentDidLoad['p-icon'], 'componentDidLoad: p-icon').toBe(2);
+    expect(status.componentDidLoad['p-button'], 'componentDidLoad: p-button').toBe(2);
 
     expect(status.componentDidUpdate['p-scroller'], 'componentDidUpdate: p-scroller').toBe(0);
     expect(status.componentDidUpdate['p-tabs-bar'], 'componentDidUpdate: p-tabs-bar').toBe(0);
 
-    expect(status.componentDidLoad.all, 'componentDidLoad: all').toBe(8);
+    expect(status.componentDidLoad.all, 'componentDidLoad: all').toBe(10);
     expect(status.componentDidUpdate.all, 'componentDidUpdate: all').toBe(0);
   });
 
@@ -342,7 +345,7 @@ test.describe('lifecycle', () => {
     expect(status.componentDidUpdate['p-tabs-item'], 'componentDidUpdate: p-tabs-item').toBe(3);
     expect(status.componentDidUpdate['p-scroller'], 'componentDidUpdate: p-scroller').toBe(0);
 
-    expect(status.componentDidLoad.all, 'componentDidLoad: all').toBe(8);
+    expect(status.componentDidLoad.all, 'componentDidLoad: all').toBe(10);
     expect(status.componentDidUpdate.all, 'componentDidUpdate: all').toBe(5);
   });
 });