+++ packages/components/src/components/link/link-styles.ts
@@ -10,6 +10,8 @@ import {
 } from '../../styles';
 import { borderRadiusSmall } from '@porsche-design-system/styles';
 
+const cssVariableInternalLinkScaling = '--p-internal-link-scaling';
+
 export const getComponentCss = (
   icon: LinkButtonIconName,
   iconSource: string,
@@ -24,7 +26,17 @@ export const getComponentCss = (
 
   return getCss(
     mergeDeep(
-      getLinkButtonStyles(icon, iconSource, variant, hideLabel, false, hasSlottedAnchor, compact, theme),
+      getLinkButtonStyles(
+        icon,
+        iconSource,
+        variant,
+        hideLabel,
+        false,
+        hasSlottedAnchor,
+        compact,
+        cssVariableInternalLinkScaling,
+        theme
+      ),
       {
         label: {
           clip: addImportantToRule('unset'), // to overrule breakpoint customizable hide-label style
@@ -56,7 +68,7 @@ export const getComponentCss = (
             '&(a)::before': {
               content: '""',
               position: 'fixed',
-              inset: '-2px',
+              inset: variant === 'ghost' ? '0px' : '-2px', // Variant ghost has no border to compensate
               borderRadius: borderRadiusSmall,
             },
             ...getFocusJssStyle(theme, { slotted: 'a', pseudo: true }),