+++ packages/components/CHANGELOG.md
@@ -14,6 +14,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 
 ### [Unreleased]
 
+#### Fixed
+
+- `Button`, `Link`: Safari rendering issue of `backdrop-filter` on border in variant `ghost`
+  ([#3435](https://github.com/porsche-design-system/porsche-design-system/pull/3435))
+
+#### Changed
+
+- `Button`, `Link`: spacings adjusted for `compact` mode
+- `Banner`, `Flyout`, `Inline Notification`, `Modal`, `Scroller`, `Toast`: button style
+  ([#3435](https://github.com/porsche-design-system/porsche-design-system/pull/3435))
+
 ### [3.17.0] - 2024-08-01
 
 ### [3.17.0-rc.2] - 2024-08-01